+++ macros/src/parse/float.rs
@@ -1,4 +1,4 @@
-use super::{common::quote_sign, int::quote_ibig};
+use super::{common::{quote_sign, quote_words}, int::quote_ibig};
 use core::str::FromStr;
 
 use dashu_base::{BitTest, Signed};
@@ -11,7 +11,7 @@ fn panic_fbig_syntax() -> ! {
     panic!("Incorrect syntax, please refer to the docs for acceptable float literal formats.")
 }
 
-pub fn parse_binary_float(embedded: bool, input: TokenStream) -> TokenStream {
+pub fn parse_binary_float(static_: bool, embedded: bool, input: TokenStream) -> TokenStream {
     let mut value_str = String::new();
     input
         .into_iter()
@@ -40,36 +40,43 @@ pub fn parse_binary_float(embedded: bool, input: TokenStream) -> TokenStream {
     assert!(signif.is_positive());
     let mag = signif.into_parts().1;
 
+    let ns = if embedded {
+        quote!(::dashu::float)
+    } else {
+        quote!(::dashu_float)
+    };
+    let repr_tt = quote!( #ns::Repr::<2> );
+    let type_tt = quote!( #ns::FBig::<#ns::round::mode::Zero, 2> );
+
     if mag.bit_len() <= 32 {
         // if the significand fits in a u32, generates const expression
         let sign = quote_sign(embedded, sign);
         let u: u32 = mag.try_into().unwrap();
 
-        if !embedded {
-            quote! {
-                ::dashu_float::FBig::<::dashu_float::round::mode::Zero, 2>
-                ::from_parts_const(#sign, #u as _, #exp, Some(#prec))
-            }
+        let value_tt = quote!( #type_tt::from_parts_const(#sign, #u as _, #exp, Some(#prec)));
+        if static_ { 
+            quote! {{ static VALUE: #type_tt = #value_tt; &VALUE }}
         } else {
-            quote! {
-                ::dashu::float::FBig::<::dashu::float::round::mode::Zero, 2>
-                ::from_parts_const(#sign, #u as _, #exp, Some(#prec))
-            }
+            value_tt
         }
     } else {
-        let signif_tt = quote_ibig(embedded, IBig::from_parts(sign, mag));
-
-        if !embedded {
+        if static_ {
+            let bytes = mag.to_le_bytes();
+            let data_defs = quote_words(&bytes, embedded);
+            let sign = quote_sign(embedded, sign);
             quote! {{
-                let repr = ::dashu_float::Repr::<2>::new(#signif_tt, #exp);
-                let context = ::dashu_float::Context::<::dashu_float::round::mode::Zero>::new(#prec);
-                ::dashu_float::FBig::from_repr(repr, context)
+                #data_defs // defines data sources
+                static VALUE: #type_tt = unsafe {
+                    #type_tt::from_repr_const(#repr_tt::from_static_words(#sign, &DATA, #exp))
+                };
+                &VALUE
             }}
         } else {
+            let signif_tt = quote_ibig(embedded, IBig::from_parts(sign, mag));
             quote! {{
-                let repr = ::dashu::float::Repr::<2>::new(#signif_tt, #exp);
-                let context = ::dashu::float::Context::<::dashu::float::round::mode::Zero>::new(#prec);
-                ::dashu::float::FBig::from_repr(repr, context)
+                let repr = #repr_tt::new(#signif_tt, #exp);
+                let context = #ns::Context::<#ns::round::mode::Zero>::new(#prec);
+                #ns::FBig::from_repr(repr, context)
             }}
         }
     }
@@ -84,33 +91,25 @@ pub fn parse_decimal_float(embedded: bool, input: TokenStream) -> TokenStream {
     let f = DBig::from_str(&value_str).unwrap_or_else(|_| panic_fbig_syntax());
     let prec = f.precision();
     let (signif, exp) = f.into_repr().into_parts();
+    
+    let ns = if embedded {
+        quote!(::dashu::float)
+    } else {
+        quote!(::dashu_float)
+    };
 
     if signif.bit_len() <= 32 {
         // if the significand fits in a u32, generates const expression
         let (sign, mag) = signif.into_parts();
         let u: u32 = mag.try_into().unwrap();
         let sign = quote_sign(embedded, sign);
-
-        if !embedded {
-            quote! { ::dashu_float::DBig::from_parts_const(#sign, #u as _, #exp, Some(#prec)) }
-        } else {
-            quote! { ::dashu::float::DBig::from_parts_const(#sign, #u as _, #exp, Some(#prec)) }
-        }
+        quote! { #ns::DBig::from_parts_const(#sign, #u as _, #exp, Some(#prec)) }
     } else {
         let signif_tt = quote_ibig(embedded, signif);
-
-        if !embedded {
-            quote! {{
-                let repr = ::dashu_float::Repr::<10>::new(#signif_tt, #exp);
-                let context = ::dashu_float::Context::new(#prec);
-                ::dashu_float::DBig::from_repr(repr, context)
-            }}
-        } else {
-            quote! {{
-                let repr = ::dashu::float::Repr::<10>::new(#signif_tt, #exp);
-                let context = ::dashu::float::Context::new(#prec);
-                ::dashu::float::DBig::from_repr(repr, context)
-            }}
-        }
+        quote! {{
+            let repr = #ns::Repr::<10>::new(#signif_tt, #exp);
+            let context = #ns::Context::new(#prec);
+            #ns::DBig::from_repr(repr, context)
+        }}
     }
 }