+++ macros/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add `static_ubig!` and `static_ibig!` macros to support static integer creation ([#38](https://github.com/cmpute/dashu/issues/38)).
+- Add `static_fbig!` macro to support static float numbers creation.
 
 ## 0.4.0